requirements.txt
asgi-redis
asgiref==3.7.2
attrs==23.1.0
autobahn==22.7.1
Automat==22.10.0
channels==4.0.0
channels-redis==4.2.0
constantly==15.1.0
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
hyperlink==21.0.0
incremental==22.10.0
msgpack-python==0.5.6
psycopg2-binary==2.9.9
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
Twisted==22.10.0
txaio==23.1.1
zope.interface==6.1
